# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Accessibility API",
    "version": "0.0.0"
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
    "schemas": {
      "Audit": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "url": {
            "type": "string",
            "format": "uri"
          },
          "type": {
            "$ref": "#/components/schemas/AuditType"
          },
          "status": {
            "$ref": "#/components/schemas/AuditStatus"
          },
          "startedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "finishedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "pagesScanned": {
            "type": "integer",
            "minimum": 0
          },
          "score": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0,
            "maximum": 100
          },
          "failureReason": {
            "$ref": "#/components/schemas/AuditFailureReason"
          }
        },
        "required": [
          "id",
          "url",
          "type",
          "status",
          "startedAt",
          "finishedAt",
          "pagesScanned",
          "score",
          "failureReason"
        ]
      },
      "AuditType": {
        "type": "string",
        "enum": [
          "free",
          "scheduled",
          "manual"
        ]
      },
      "AuditStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "completed",
          "failed"
        ]
      },
      "AuditFailureReason": {
        "type": [
          "string",
          "null"
        ],
        "enum": [
          "forbidden_url",
          "robots_disallowed",
          "scan_failed",
          null
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "error",
          "message"
        ]
      },
      "FreeAuditRequest": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "format": "uri"
          }
        },
        "required": [
          "url"
        ]
      },
      "AuditReport": {
        "type": "object",
        "properties": {
          "audit": {
            "$ref": "#/components/schemas/Audit"
          },
          "totalIssues": {
            "type": "integer",
            "minimum": 0
          },
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReportGroup"
            }
          },
          "automatedCoverageNotice": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "audit",
          "totalIssues",
          "groups",
          "automatedCoverageNotice"
        ]
      },
      "ReportGroup": {
        "type": "object",
        "properties": {
          "ruleId": {
            "type": "string",
            "minLength": 1
          },
          "title": {
            "type": "string"
          },
          "helpUrl": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "impact": {
            "$ref": "#/components/schemas/Impact"
          },
          "occurrences": {
            "type": "integer",
            "minimum": 1
          },
          "examples": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReportExample"
            }
          },
          "wcagCriteria": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "rgaaCriteria": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "ruleId",
          "title",
          "helpUrl",
          "impact",
          "occurrences",
          "examples",
          "wcagCriteria",
          "rgaaCriteria"
        ]
      },
      "Impact": {
        "type": "string",
        "enum": [
          "minor",
          "moderate",
          "serious",
          "critical"
        ]
      },
      "ReportExample": {
        "type": "object",
        "properties": {
          "selector": {
            "type": "string"
          },
          "htmlExcerpt": {
            "type": "string"
          }
        },
        "required": [
          "selector",
          "htmlExcerpt"
        ]
      },
...
      "ManualCheckList": {
        "type": "object",
        "properties": {
          "referentialVersion": {
            "type": "string"
          },
          "themes": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "number": {
                  "type": "integer"
                },
                "title": {
                  "type": "string"
                }
              },
              "required": [
                "number",
                "title"
              ]
            }
          },
          "criteria": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "theme": {
                  "type": "integer"
                },
                "title": {
                  "type": "string"
                },
                "autoTested": {
                  "type": "boolean"
                },
                "axeRules": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "autoProblems": {
                  "type": "integer",
                  "minimum": 0
                },
                "check": {
                  "type": [
                    "object",
                    "null"
                  ],
                  "properties": {
                    "status": {
                      "$ref": "#/components/schemas/ManualStatus"
                    },
                    "notes": {
                      "type": "string"
                    },
                    "evidenceUrl": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "checkedBy": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "checkedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "status",
                    "notes",
                    "evidenceUrl",
                    "checkedBy",
                    "checkedAt"
                  ]
                }
              },
              "required": [
                "id",
                "theme",
                "title",
                "autoTested",
                "axeRules",
                "autoProblems",
                "check"
              ]
            }
          },
          "progress": {
            "type": "object",
            "properties": {
              "checked": {
                "type": "integer",
                "minimum": 0
              },
              "total": {
                "type": "integer",
                "minimum": 0
              }
            },
            "required": [
              "checked",
              "total"
            ]
          }
        },
        "required": [
          "referentialVersion",
          "themes",
          "criteria",
          "progress"
        ]
      },
      "ManualStatus": {
        "type": "string",
        "enum": [
          "ok",
          "ko",
          "na"
        ]
      },
      "UpsertManualCheckRequest": {
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/ManualStatus"
          },
          "notes": {
            "type": "string",
            "maxLength": 5000
          },
          "evidenceUrl": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 2000,
            "format": "uri"
          }
        },
        "required": [
          "status"
        ]
      },
...
    },
...
  },
  "paths": {
    "/audits/free": {
      "post": {
        "summary": "Start a free audit of a public URL",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FreeAuditRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Audit queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Audit"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/audits/{id}": {
      "get": {
        "summary": "Get the status of an audit",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Audit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Audit"
                }
              }
            }
          },
          "404": {
            "description": "Audit not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/audits/{id}/report": {
      "get": {
        "summary": "Get the report of a completed audit",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Report",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditReport"
                }
              }
            }
          },
          "404": {
            "description": "Audit not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Audit not completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/sites/{id}/manual-checks": {
      "get": {
        "summary": "RGAA criteria with what was verified by hand and what the scan sees",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Criteria",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManualCheckList"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/sites/{id}/manual-checks/{criterionId}": {
      "put": {
        "summary": "Record the manual verification of one criterion",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d{1,2}\\.\\d{1,2}$"
            },
            "required": true,
            "name": "criterionId",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertManualCheckRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Recorded"
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Forget the manual verification of one criterion",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d{1,2}\\.\\d{1,2}$"
            },
            "required": true,
            "name": "criterionId",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Removed"
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}